general cleanup/rename to lispdocs
